@@ -1,9 +1,10 @@
-import { OnInit } from '@angular/core';
+import { OnInit, Directive } from '@angular/core';
 import { BaseDataEntryComponent } from 'imng-kendo-data-entry';
 
 import { BuildingCrudFacade } from './crud.facade';
 import { IBuilding, BuildingProperties, BuildingFormGroupFac } from '../../../models/units-odata';
 
+@Directive()
 export abstract class BuildingBaseEntryComponent extends BaseDataEntryComponent<IBuilding, BuildingCrudFacade>
   implements OnInit {
   public readonly props = BuildingProperties;
