@@ -1,9 +1,10 @@
-import { OnInit } from '@angular/core';
+import { OnInit, Directive } from '@angular/core';
 import { BaseDataEntryComponent } from 'imng-kendo-data-entry';
 
 import { CertificationCrudFacade } from './crud.facade';
 import { ICertification, CertificationProperties, CertificationFormGroupFac } from '../../../models/certifications-odata';
 
+@Directive()
 export abstract class CertificationBaseEntryComponent extends BaseDataEntryComponent<ICertification, CertificationCrudFacade>
   implements OnInit {
   public readonly props = CertificationProperties;
