@@ -1,10 +1,11 @@
 import { isPlatformBrowser } from '@angular/common';
 import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
-import { Log, OidcClient, SigninRequest, SignoutRequest, UserManager } from 'oidc-client';
-import { from, Observable } from 'rxjs';
+import { Log, OidcClient, SigninRequest, SignoutRequest, UserManager, UserManagerSettings } from 'oidc-client';
+import { from, Observable, of } from 'rxjs';
 import { Config, OIDC_CONFIG } from '../models/config.model';
 import { OidcEvent, StorageKeys } from '../models/constants';
 import { IOidcUser } from '../models/oidc-user';
+import { catchError } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root',
@@ -12,6 +13,7 @@ import { IOidcUser } from '../models/oidc-user';
 export class OidcService {
   private readonly _oidcUserManager: UserManager;
   private readonly _oidcClient: OidcClient;
+  private readonly _userManagerSettings: UserManagerSettings;
 
   private readonly _useCallbackFlag: boolean = true;
 
@@ -36,7 +38,7 @@ export class OidcService {
         userStore: clientSettings.userStore,
       };
     }
-
+    this._userManagerSettings = { ...clientSettings };
     this._oidcUserManager = new UserManager(clientSettings);
     this._oidcClient = new OidcClient(clientSettings);
   }
@@ -133,8 +135,10 @@ export class OidcService {
 
   signOutRedirect(args?: any): Observable<any> {
     this.setCallbackInformation(false);
-
-    return from(this._oidcUserManager.signoutRedirect({ ...args }));
+    return from(this._oidcUserManager.signoutRedirect({ ...args })).pipe(catchError(err => {
+      console.error(err);
+      return of();
+    }));
   }
 
   signInSilent(args?: any): Observable<IOidcUser> {
@@ -168,7 +172,8 @@ export class OidcService {
   private setCallbackInformation(isPopupCallback: boolean) {
     // is browser and useCallbackFlag set to true or defaults to true
     if (isPlatformBrowser(this.platformId) && this._useCallbackFlag) {
-      sessionStorage.setItem(StorageKeys.PopupCallback, `${isPopupCallback}`);
+      localStorage.setItem(StorageKeys.PopupCallback, `${isPopupCallback}`);
+      localStorage.setItem(StorageKeys.OidcSettings, JSON.stringify(this._userManagerSettings));
     }
   }
 }
